+++ backend-rust/src/migrations.rs
@@ -11,6 +11,7 @@ mod m0005_fix_dangling_delegators;
 mod m0006_fix_stake;
 mod m0008_canonical_address_and_transaction_search_index;
 mod m0010_fill_capital_bound_and_leverage_bound;
+mod m0014_baker_metrics;
 
 /// Ensure the current database schema version is compatible with the supported
 /// schema version.
@@ -203,14 +204,16 @@ pub enum SchemaVersion {
     TrackRemovedBakers,
     #[display("0013:Fix delegated_restake_earnings data in accounts")]
     FixDelegatedStakeEarnings,
+    #[display("0014:RankingByLotteryPower")]
+    BakerMetrics,
 }
 impl SchemaVersion {
     /// The minimum supported database schema version for the API.
     /// Fails at startup if any breaking database schema versions have been
     /// introduced since this version.
-    pub const API_SUPPORTED_SCHEMA_VERSION: SchemaVersion = SchemaVersion::TrackRemovedBakers;
+    pub const API_SUPPORTED_SCHEMA_VERSION: SchemaVersion = SchemaVersion::BakerMetrics;
     /// The latest known version of the schema.
-    const LATEST: SchemaVersion = SchemaVersion::FixDelegatedStakeEarnings;
+    const LATEST: SchemaVersion = SchemaVersion::BakerMetrics;
 
     /// Parse version number into a database schema version.
     /// None if the version is unknown.
@@ -239,6 +242,7 @@ impl SchemaVersion {
             SchemaVersion::RankingByLotteryPower => false,
             SchemaVersion::TrackRemovedBakers => false,
             SchemaVersion::FixDelegatedStakeEarnings => false,
+            SchemaVersion::BakerMetrics => false,
         }
     }
 
@@ -330,7 +334,11 @@ impl SchemaVersion {
                     .await?;
                 SchemaVersion::FixDelegatedStakeEarnings
             }
-            SchemaVersion::FixDelegatedStakeEarnings => unimplemented!(
+            SchemaVersion::FixDelegatedStakeEarnings => {
+                let next_schema_version = SchemaVersion::BakerMetrics;
+                m0014_baker_metrics::run(&mut tx, endpoints, next_schema_version).await?
+            }
+            SchemaVersion::BakerMetrics => unimplemented!(
                 "No migration implemented for database schema version {}",
                 self.as_i64()
             ),