+++ backend-rust/CHANGELOG.md
@@ -8,7 +8,18 @@ All notable changes to this project will be documented in this file.
 
 - Support for `Query::bakers` sorting by block commission.
 
-## [0.1.34] - 2025-03-06
+## [0.1.35] - 2025-03-14
+
+Database schema version: 14
+
+### Added
+
+- Add baker metrics endpoints
+- Gathering statistics about change in amount of bakers
+- Migrate bakers statistics
+- Indexing genesis block creates baker metrics
+
+## [0.1.34] - 2025-03-13
 
 Database schema version: 13